Move processing of tokens for print into GreekText.
This pre-processing, (like prepending spaces) was done
along with colorizing inside get_colorized_text_for_tokens.
Print-preprocessing is now a separate step done in GreekText
(greek_text.py), and the resulting list of printable tokens
is exposed as a new printable_tokens property on GreekText.

This refactor supports the next step of building an HTML renderer.